Set up search output page and reorganized search components - having error capturing what was submitted in the seearch bard and also rendering this results in the page
