+++ lyrios-frontend/src/components/VideoOutput.js
@@ -59,6 +59,7 @@ function VideoOutput (props) {
             {youtubeData ?   <>     
             {youtubeData.map((value, index)=>{
                 console.log(value)
+                
                 return(
                     <div className = 'videoItem'>
                        <h4 className = 'videoTitle'>{value.snippet.title}</h4>  <YouTube videoId= {value.id.videoId} className = 'videoIndex'/>
